lab/requirements.txt
# Core Data Science
pandas==2.2.0
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# HTTP & APIs
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data Processing
pyarrow==14.0.1
openpyxl>=3.0.0,<4.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-xdist==3.5.0

# Code Quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Logging & Monitoring
python-json-logger==3.2.1

# Jupyter
jupyter==1.0.0
jupyterlab==4.0.9
ipython==8.18.1

# Utilities
tqdm==4.66.1
click==8.1.7
pyyaml==6.0.1

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# OmegaSports Engine (when available)
# omega-sports-engine>=1.0.0

